Switched to using FMDatabasePool for song model reads
Contrary to what I for some reason assumed, FMDatabase connections CANNOT be used concurrently. For concurrent reads, an FMDatabasePool must be used to aquire an available connection.

Notmally you would not use a pool for concurrency on a read/write database because you CANNOT read and write concurrently in a pool without deadlocking. However, since we're using WAL mode thus allowing concurrent reads with a single write and we use an FMDatabaseQueue to serialize writes, we can use the pool successfully for a read/write database.